+++ docs/md/releasenotes.md
@@ -8,6 +8,8 @@
 
 - Fix side effect on `DetailedArrayReferenceValue` modifying values at previous offsets on array store
   instructions during `PartialEvaluator` execution.
+- Fix `JvmTransferRelation` to produce a successor in case of missing interprocedural call edge
+  (e.g., in case of incomplete call resolution)
 
 ## Version 9.0.3